added ts & prettier
