+++ package.json
@@ -5,7 +5,7 @@
     "type": "module",
     "scripts": {
         "dev": "vite",
-        "build": "vite build",
+        "build": "tsc && vite build",
         "preview": "vite preview"
     },
     "dependencies": {
@@ -14,12 +14,15 @@
         "react-router-dom": "^6.4.1"
     },
     "devDependencies": {
+        "@types/node": "18.8.2",
         "@types/react": "^18.0.17",
         "@types/react-dom": "^18.0.6",
         "@vitejs/plugin-react": "^2.1.0",
         "autoprefixer": "^10.4.11",
         "postcss": "^8.4.16",
+        "prettier": "^2.7.1",
         "tailwindcss": "^3.1.8",
+        "typescript": "^4.6.4",
         "vite": "^3.1.0"
     }
 }
