--- src/components/ErrorScreen.jsx
@@ -1,7 +1,0 @@
-export default function ErrorScreen() {
-    return (
-        <div className="h-screen flex flex-col items-center justify-center text-red-600 text-3xl">
-            server not working!
-        </div>
-    );
-}
