--- src/components/LoadingScreen.jsx
@@ -1,7 +1,0 @@
-export default function LoadingScreen() {
-    return (
-        <div className="h-screen flex flex-col items-center justify-center text-3xl">
-            i'm loading bruh
-        </div>
-    );
-}
