+++ src/components/LoadingScreen.tsx
@@ -1,0 +1,7 @@
+export default function LoadingScreen() {
+    return (
+        <div className="h-screen flex flex-col items-center justify-center text-3xl">
+            i'm loading bruh
+        </div>
+    );
+}
